changed to double - hot fix to prd

+++ demo/src/main/java/com/lbd/demo/Calculator.java
@@ -2,11 +2,12 @@ package com.lbd.demo;
 
 public class Calculator {
 
-	public int add(int a, int b){
+	public double add(double a, double b){
 		return a+b;
 	}
 	
-	public int substract(int a, int b){
+	public double substract(double a, double b){
 		return a-b;
 	}
+	
 }
